move request to a separate file, create queue wrapper, fix checking numbers in progress
