+++ crates/net/downloaders/src/bodies/mod.rs
@@ -1,2 +1,5 @@
 /// A naive concurrent downloader.
 pub mod concurrent;
+
+mod queue;
+mod request;
